gpexpand: Fix tables not copied to new segments
**Issue:**
A customer reported an issue where they were unable to DROP time-based login constraints for users after running gpexpand. This was because, if there are time-based login constraints for users, there are entries in the `pg_auth_time_constraints` catalog table that are identical across all segments. However, when gpexpand was run, this table was not replicated to new segments. As a result, running the DROP command would give an error.
```
NOTICE:  dropping DENY rule for "user03" between Thursday 18:00:00 and Saturday 23:59:59
ERROR:  cannot find matching DENY rules for "user03"  (seg3 192.168.50.53:6000 pid=4429)
```

**RCA:**
The reason why the table was not getting replicated is because gpexpand creates a template of the coordinator segment and uses it to create new segments in the cluster. As part of the cleanup process(since it uses coordinator template), some catalog tables present only on the coordinator are being removed from the new segments. This list of tables is defined in the `COORDINATOR_ONLY_TABLES` list in `gpcatalog.py`.

Catalog tables like `pg_auth_time_constraint`, `pg_description`, and `pg_shdescription` are also part of the `COORDINATOR_ONLY_TABLES` list, but these are present on all segments and not just on the coordinator.

**Fix:**
This commit removes the tables mentioned above from the `COORDINATOR_ONLY_TABLES` list.

Also added a behave test case to verify if these tables are present on all the newly created segments.