Add filenum field to pg_attribute_encoding
For CO tables, today, each column is tied to an implicit "filenum",
which is a shorthand for the file range that the column is assigned.
This means segfiles for columns are of the form:
(rfnode = relfilenode)

attnum 1: rfnode, rfnode.1 .. rfnode.127 (implicit filenum = 1)
attnum 2: rfnode.128, rfnode.129 .. rfnode.255 (implicit filenum = 2)
attnum 3: rfnode.256, rfnode.257 .. rfnode.383 (implicit filenum = 3)
...

The range of relfiles for a column depends on this implicit "filenum",
which is always equal to the attnum.

We want to decouple the implicit filenum from the attnum. To illustrate,
segfiles can now be of the form:

attnum 1: rfnode.256, rfnode.257 .. rfnode.383 (explicit filenum = 3)
attnum 2: rfnode, rfnode.1 .. rfnode.127 (explicit filenum = 1)
attnum 3: rfnode.128, rfnode.129 .. rfnode.255 (explicit filenum = 2)
...

This commit does exactly that by introducing
pg_attribute_encoding.filenum, so we can have a mapping
between attnum and filenum.

We will leverage this decoupling in future commits.
For instance, to support column-only rewrites,
we would need a way to create a "rewritten" relfile
for an existing column i and then swap them in the catalog,
while supporting MVCC semantics etc.

It is also now worth observing that the filenum no longer needs to
range from 1..1600 (MaxHeapAttributeNumber).
It can now go beyond 1600, with more alter operations.

Also, we have edited the tests to include the new filenum field.
Some tests don't need to check filenum,
but keeping the field to check if further commits/changes affect it