Fix error in ATSETAM to ao_column with dropped column
When we alter the table AM from non-AOCO to AOCO, we need to add
pg_attribute_encoding entries for it. That includes checking types
of the columns to decide the encoding options. But if a column is
dropped, we would fail to find its type. Now take that into
consideration. Basically a dropped column will still have its entry
in pg_attribute_encoding (if it's an AOCO table), until the table
itself is dropped, just like pg_attribute.