Double gpfdist listening to one port
In an extreme case, two gpfdist occupy one port(ipv4/ipv6). The reason for this problem includes two aspects. The first one is that 'bind' is not mutually exclusive. And another is that when listening to a port fail, gpfdist will try the same port with a different protocol(ipv4 or ipv6).

So the PR fixes the problem by changing the handling for failed listening behavior.

(cherry picked from commit d58fde9741950fca532455bd222979c93219bbb5)