Resolve merge fixmes from syncrep.c
The implementation questioned is not new and has been there for
GPDB4,5,6 and 7. Longer term goal is to make it better but nothing in
short term to fix here.

Plus, the fixme in SyncRepGetSyncStandbys() can't be done as
standbconfig is not created or populated for
coordinator-standby. Hence, the code needs to live in
SyncRepGetSyncStandbys(). Once proper implementation is done for
coodinator and standby then this part can be tackled as well.

(cherry picked from commit 416351029259afe37c1009235a5d870d9b1ab289)