brin ao/co: Fix final partial range summarization
This commit fixes partial range summarization for the final range in AO
segments, in a given AO/CO table.

(1) Earlier, we incorrectly used to call RelationGetNumberOfBlocks() in
summarize_range(). While that is okay for heap tables, it isn't for
AO/CO tables. What we need is a way to get at the BlockSequence in which
a given heap block lies. Once we do so, we can easily get the upper
bound on the number of logical heap blocks in that segment. We have
introduced the table AM API relation_get_block_sequence() to this end.

(2) We also make the decision to avoid final partial range summarization
in case the final partial range was extended underneath us by a
different insert session. This is done by checking if the number of
blocks in the segment is still the same, as what we determined at the
beginning of summarization, and bailing accordingly.

(3) The assert inside summarize_range() only holds true for heap tables:
heapBlk % state->bs_pagesPerRange == 0
It wasn't exposed earlier due since no tests had pages_per_range > 1.
Relax accordingly.

(4) Added FastSequenceGetNumHeapBlocks() to extract the logic to obtain
number of logical heap blocks from a given lastSequence value. Also
added AOSegment_PopulateBlockSequence() to populate a given
BlockSequence, provided with a segrelid and segno.

Note: relation_get_block_sequence() will be useful for specific range
summarization in the future.

(cherry picked from commit ca9379e4541da85b0c02ddbd90862fc5fbbb9109)