Optimize retrieving relation columns in Orca
In 7X, we need to retrieve each individual partition of a partitioned
table, which makes the logic to retrieve relation columns a new hot spot
in the code and taking up 60-70% of the optimization time for large and
wide partitioned tables.

Previously, we would attempt to retrieve the attribute statistics in
order to get the column width. In cases where this is a fixed width
(ints, floats, dates, etc.), we can get this from pg_attribute itself
which is much cheaper.

As an example, for a 1200 partition, 200 column table of type `int` with
no dropped/swapped columns, a simple `select * from part_table` query
went from 3.5s to 1.1s with an empty relcache

Also, call GetAttAvgWidth instead of GetAttStats to calculate the width
if necessary. This is slightly faster for non fixed width columns like
varchar, for example 1.6s instead of 2s if all columns above are
of type varchar.

(cherry picked from commit effe9045a040251c32c27a4b98bf0361a2c63d6a)