remove resGroupId check in runaway detector
In previous version, we disable red-zone check for resource group,
the resgroup use cgroup to clean the process which consumes memory
exceeds the rg's configuration.

For the new version for resgroup, we use redZoneChunks to determine
whether the current process is in red-zone or not, so remove the
resGroupId check.

(cherry picked from commit 15a9683ee320c4ccc64d785d7ebebb28529552ed)