Fix duplicate filter due to operator argument order (#15111)
Prior to commit 063b0abc3c, ORCA preprocessor had an assumption that
after normalization all duplicate predicates would be in the form:

  (IDENT op CONST)

However, after that commit the predicate can now also be in form:

  (CONST op IDENT)

Issue is that ORCA did not consider when the above two forms could be
duplicates of each other. If the types on either side of the compare are
different then the operator id is also likely different. In that case,
ORCA could not detect the commutative equality and thus created a
duplicate filter.

Following is minimal example that demonstrated the issue:

  CREATE TABLE t1 (id integer);
  CREATE TABLE t2 (id smallint);

  EXPLAIN (costs off) SELECT * FROM t1 a JOIN t2 fa ON a.id = fa.id WHERE a.id = 1;

After normalization and predicate push down the tree looks like:

  +--CLogicalNAryJoin
     |--CLogicalSelect
     |  |--CLogicalGet "t1" ("t1"), Columns: ["id" (0), ...
     |  +--CScalarBoolOp (EboolopAnd)
     |     |--CScalarCmp (=)(0.96.1.0)
     |     |  |--CScalarIdent "id" (0)
     |     |  +--CScalarConst (1)
     |     +--CScalarCmp (=)(0.96.1.0)
     |        |--CScalarIdent "id" (0)
     |        +--CScalarConst (1)
     |--CLogicalSelect
     |  |--CLogicalGet "t2" ("t2"), Columns: ["id" (8), ...
     |  +--CScalarBoolOp (EboolopAnd)
     |     |--CScalarCmp (=)(0.533.1.0)
     |     |  |--CScalarConst (1)
     |     |  +--CScalarIdent "id" (8)
     |     +--CScalarCmp (=)(0.532.1.0)
     |        |--CScalarIdent "id" (8)
     |        +--CScalarConst (1)
     +--CScalarBoolOp (EboolopAnd)
        |--CScalarCmp (=)(0.533.1.0)
        |  |--CScalarIdent "id" (0)
        |  +--CScalarIdent "id" (8)
        +--CScalarCmp (=)(0.533.1.0)
           |--CScalarIdent "id" (0)
           +--CScalarIdent "id" (8)

Notice that the CScalarCmp nodes above t2 use operator ids 533 and 532
which are commutative equal operators of each other, which ORCA did not
know how to deduplicate.

Fix is for predicate deduplicator to consider commutative equality
operators. This information exists in the pg_operator catalog as oprcom
and already exists in ORCA metadata accessor.

(cherry picked from commit c3c0aef40158bb5f42c5dc5bbbf21aac11cd7f3f)