Allow queries with valid function variadic flag
Through this PR, we have allowed processing of functional Expressions with 'funcvariadic' as true.

Prior to this update, if for any functional expression 'funcvariadic' flag
 was 'true' then ORCA would fall back to planner.
An exception was raised in CTranslatorScalarToDXL::TranslateFuncExprToDXL(,).

To solve the problem, we passed the 'funcvariadic' through to the planned statement.

Two approaches were tried,

APPROACH 1- Used the data coming in ‘func_expr->funcvariadic’ (at Query to DXL creation)
as the base value and passed it all through out ORCA.

APPROACH 2- Add the 'funcvariadic' flag value in CMDFunctionGPDB (from pg_proc) and
used it from here. This approach was tried in following PR (closed) #15236

APPROACH-1 was selected because as we tried to read the 'funcvariadic' from the PG_PROC catalog
instead of the value we receive after parsing the query. It was found that using APPROACH-2,
for some queries there was a difference in the value read from both the points, this caused an error.
Eg: select format('|%20s|', 'ten');

'funcvariadic' flag from query - FALSE
'funcvariadic' flag from Catalog - TRUE

This caused failure in "Datum text_format(PG_FUNCTION_ARGS)"
Thus, it was decided not to read value from catalog and continue with the value received after parsing the query.

(cherry picked from commit 71bf73bee505d2518650b61142b67fc613254c35)