Disable GUC optimizer_penalize_broadcast_threshold when set to 0 (#15209)
Currently, when optimizer_penalize_broadcast_threshold is set to 0, it
will disable broadcasts when the estimated tuples to be broadcast is
greater than 0. Because the value of this GUC is limited to INT_MAX,
there is no way to disable this guc. This commit disables penalizing
broadcast when this value is 0.

(cherry picked from commit 525b87facb2f9a806386f9a5de53a63102e1735b)