Fix: 'unrecognized node type: 145' in transformExpr (#15184)
In GPDB QD node, the execution path is the same as PG12:
```T_TypeCast -> T_TypeCast -> T_BooleanTest -> T_A_EXPr```
However, after transformExpr(pstate, raw_default,......) in QD, we changed the path of raw_default to
```T_TypeCast -> T_TypeCast -> T_BooleanTest -> T_OpExpr```

Then we dispatch it (alterStmt) to QE, we need to call transformExpr() in QE too. But the path of raw_default is changed. And we cannot process OpExpr in transformExpr since it's not the structure of a parse tree.

So the fix is easy: we have to let transformBooleanTest not change its parameter.
Details please see the discussion in PR threads.