fix tids are not in order when building bitmap index (#15271)
When we build a bitmap index on a heap table, we need to scan all the
tuples. Usually, it'll start at block 0 to the end. But if we allow sync scan,
it'll scan the table circularly, from block X up to the end and then from
block 0 to X-1, to ensure we visit all rows while still participating in the
common scan. This could lead tids are not in order when building bitmap
index, which will throw an error in the current implementation.

So this PR forbids sync scan when building bitmap index.

Besides the above fact, if we build bitmap index on a heap table, there could
have HOT-chain, and it'll return the root tuple's tid, which could lead to the
TIDs we scanned are not in order.

(cherry picked from commit 68c758bd4394c6c16563e1e69e4bf1785b5e04cc)