Update Array Coerce Cast Metadata object
For certain queries, ORCA generates an Array Coerce Expr. To correctly
generate this, in addition to the current information that is passed as
part of ArrayCoerceCast MD object, we also require the element type.
This commit updates that information and generates the equivalent
ArrayCoerceExpr.

(cherry picked from commit 228d19ac77d33b010793d594cc12209e7620dc66)