[ORCA] Support boolean static partition pruning  (#15348)
After Postgres 12 merge into GPDB, GPDB absorbed the underlying Postgres
partition implementation. In the process, ORCA lost static partition
pruning on boolean partitioned tables. In order to support that feature,
ORCA constraint framework needs to support boolean types as interval
constraints for constraint contradiction.

In SQL standard, boolean type includes NULL and UNKNOWN. Some special
care was needed to handle UNKNOWN type as it relates to an interval
constraint. Since it is unknown, it can be anywhere on the spectrum
(-inf, inf), hence that is the range we assign it.

In the process of supporting this feature, a related issue was
discovered in upstream Postgres [1].

This commit also enables partition pruning with general boolean compare.
It is important to note that general comparison handles NULL different
than boolean test. For example, this evaluates to TRUE:

    SELECT true IS NOT null;

However, this evaluates NULL:

    SELECT true<>null;

This has vastly different effects on static partition pruning. First
step to enabling this general boolean compare is to create interval
constraints from boolean scalar idents (see PciIntervalFromScalarIdent).
Second step is to update the interval range to support domain of boolean
types (see CConstraintInterval::PciUnbounded). In order to illustrate
that issue, consider the following table:

    CREATE TABLE booltab (a bool) partition by list(a) (
      partition part_t values(true),
      partition part_f values(false),
      partition default def);

The default partition constraint is defined in the catalog. And in ORCA
world is represented as an expression like:

    +--CScalarBoolOp (EboolopNot)
       +--CScalarBoolOp (EboolopAnd)
          |--CScalarBoolOp (EboolopNot)
          |  +--CScalarNullTest
          |     +--CScalarIdent "a" (0)
          +--CScalarArrayCmp Any (=)
             |--CScalarIdent "a" (0)
             +--CScalarArray: {(0) (1)}

If we track the constraint derivation at each level without adjusting
the constraint interval for boolean, then the derivation looks like:

    +--CScalarBoolOp (EboolopNot)           {"a" (0), ranges: (-inf, 0) (0, 1) (1, inf) [NULL] }
       +--CScalarBoolOp (EboolopAnd)        {"a" (0), ranges: [0, 0] [1, 1] }
          |--CScalarBoolOp (EboolopNot)     {"a" (0), ranges: (-inf, inf) }
          |  +--CScalarNullTest             {"a" (0), ranges: [NULL] }
          |     +--CScalarIdent "a" (0)
          +--CScalarArrayCmp Any (=)        {"a" (0), ranges: [0, 0] [1, 1] }
             |--CScalarIdent "a" (0)
             +--CScalarArray: {(0) (1)}

When the predicate is a=false, then the constraint is:

    +--CScalarBoolOp (EboolopNot)           {"a" (0), ranges: (-inf, 1) (1, inf) }
       +--CScalarIdent "a" (1)              {"a" (0), ranges: [1, 1] [NULL] }

The conjunction of table constraint:

    {"a" (0), ranges: (-inf, 0) (0, 1) (1, inf) [NULL] }

And predicate constraint:

    {"a" (0), ranges: (-inf, 1) (1, inf) }

Produces result constraint:

    {"a" (0), ranges: (-inf, 0) (1, inf) }

The problem with that result constraint is that it is not considered a
contradiction because there is a range, albeit outside the range of
boolean [0,1]. As a result, the default partition is incorrectly not
pruned.

There are 2 different ways to solve this:
   1) Fix up contradiction detection to special case boolean intervals
   2) Fix up boolean interval to map to the (-inf, inf) domain space

One problem with the 2nd approach is that other systems (e.g. direct
dispatch) use the constraint system and rely on finite values (i.e. 0,
1). That would mean fixing up those areas in the code as well to handle
the full spectrum (-inf, inf) correctly.

Instead, this commit implements the 1st approach. It also requires
updating the EboolopNot operator to keep the domain [0, 1]. In this case
the above default partition and predicate constraint derivation looks
like:

    +--CScalarBoolOp (EboolopNot)           {"a" (0), ranges: [NULL] }
       +--CScalarBoolOp (EboolopAnd)        {"a" (0), ranges: [0, 0] [1, 1] }
          |--CScalarBoolOp (EboolopNot)     {"a" (0), ranges: [0, 0] [1, 1] }
          |  +--CScalarNullTest             {"a" (0), ranges: [NULL] }
          |     +--CScalarIdent "a" (0)
          +--CScalarArrayCmp Any (=)        {"a" (0), ranges: [0, 0] [1, 1] }
             |--CScalarIdent "a" (0)
             +--CScalarArray: {(0) (1)}

    +--CScalarBoolOp (EboolopNot)           {"a" (0), ranges: [0, 0] }
       +--CScalarIdent "a" (1)              {"a" (0), ranges: [1, 1] [NULL] }

Then the conjunction of the two produces:

    {"a" (0), ranges: }

And the table is correctly pruned.

[1] https://www.postgresql.org/message-id/flat/CAHnPFjQ5qxs6J_p+g8=ww7GQvfn71_JE+Tygj0S7RdRci1uwPw@mail.gmail.com

(cherry picked from commit ebccd1f8806d6bedd6e4f127505b1e62f4a8789f)