get correlation from segments instead of calculating it on QD (#15357)
We cannot use the same method as PostgreSQL does to calculate the correlation in QD.
When we collect data from segments to QD, this will change the physical order of the data.
such as in segment 1 the data is 1,3,5,7,9. And in segment 2 the data is 2,4,6,8,10.
In each segment the data is ordered, and correlation is 1 in each segment.
But after we collect the data to QD, it may be 1,3,5,2,4,7,9,6,8,10. And the correlation
is 0.3 or something else and it is not stable.
And this will increase the cost of index scan which is shouldn't be done.

So get correlations from segments and then calculate correlation for QD.

we use the weighted mean algorithm to calculate correlation on QD,
However, In some situations, we may not be able to obtain reltuples of a table,
such as none-leaf part of partitioned table or the parent table of the
inherited table. So we can only use the mean algorithm to calculate correlation for these tables.

(cherry picked from commit 69ef4dad20c2981d3c87e58648047f97a8d5d1b4)