Using __gp_aoseg/__gp_aocsseg in missing/orphaned file views
The __gp_aoseg/__gp_aocsseg functions provide more details such as the eof
of segments. Use them for the check missing/orphaned file views. So for
the missing file view, ignore those w/ eof<=0. They could be present just
for the bookkeeping of modcount. For example, at the end of DELETE, when
we are increasing modcount, we choose any non-reserved segment to do that.
So say if we've only inserted into the reserved segment 0 before, then we'll
choose segment 1 to update modcount. That will create the entry for segno=1
in aoseg table, but we do not really insert anything to the data file so it
doesn't exist.

(cherry picked from commit 623db83b8ea80cfe47627563d45e107b5c175b2e)