Improve refresh materialized view with "no data" option
In postgres, with option "WITH NO DATA", the materialized view will be
truncated without run the query.

In GPDB, the refresh clause is dispatched to segments when execute the
query plan. But for WITH NO DATA option, it should be like a TRUNCATE
like in postgres, so it doesn't need to take long time to run the query.

By adding an constant-FALSE to the qual to generate a result plan with
One-Time Filter. We have hold matview's distribution policy into the
query->intoPolicy, planner will create a motion node on the top so that
the refresh clause will still dispatch to segments with the query plan
and the query can be execute quickly.

(cherry picked from commit ec848dee4ac35157a90f9e186ddcdcd4a72458be)