brin: Remove upper pages for AO/CO tables
This commit removes the upper page architecture for AO/CO tables. This
is in preparation of a rewrite of BRIN for AO/CO tables.

The objective of this commit is to align the existing BRIN code as much
as possible with upstream, and to ensure that tests with BRIN indexes on
heap tables pass. Existing AO/CO tests will fail.

Key points:

(1) Reverted revmap_physical_extend to not returning a BlockNumber. It
was a dead return anyway.

(2) Reverted skipExtend parameter for brin_doupdate(). Its not clear why
this was necessary. It shouldn't be necessary for heap tables.

(3) Kept isAo parameter for metapage and xlog.

(cherry picked from commit 4050cd74436ba47fae4f1b457e5bcfcfc79f48c7)