Update distribution policy for dropping distribution key dependency
Prior to this commit, if DROP TYPE..CASCADE finds a distribution key
column dependent on the type being dropped it would go ahead and drop
the column without updating catalog for table's distribution policy
leading to invalid distribution policy in gp_distribution_policy and
when trying to describe the table it would refernce dropped distribution
column that would lead to a crash. To avoid landing in this state, this
commit aligns the behavior of DROP TYPE..CASCADE with that of a direct
drop (ALTER ... DROP COLUMN) for a distribution key column and update
the distribution policy of the table to random.  This commit adds tests
and distributions policy asserts for few existing tests.

(cherry picked from commit 41e8e99e49e3c55d481778d26b64894124f4ae2d)