Fix an issue that we are not checking file .204800 in ao_foreach_extent_file
This caused an issue that we are not dropping the extension file .204800
which is a valid file that can be created after a series of AT command (i.e.
rewriting the first column in the table that has segno=0). Found in CI test
when we check gp_toolkit.gp_check_orphaned_files.

(cherry picked from commit 7d0cf0e5da1fce3d96ba3eb89b5edbddc565586b)