Fix an issue where we increment command counter incorrectly
In commit d3dcb9bc97d8 we added an CCI in swap_relation_files() after
we swapped AM. But that change should really be limited to AM change and
nothing else.

This is revealed by the previous commit where we allowed TAP test to
run vacuum successfully. So, when we are running VACUUM FULL pg_attribute,
we could have an errorneous state where the rebuilt pg_attribute does not
contain the entries for the intermediate table whereas we are building table
descriptor for the intermediate table as part of the CCI which requires
pg_attribute entries for the intermediate table... In overall I think CCI
during swapping relation should be discouraged but if not avoidable at least
we shouldn't do that for critical catalogs.

(cherry picked from commit c0124395eac2a0437609a4eaf92351d5ae10bf73)