Move a FIXME from postgresql.conf.sample to guc.c
The sample configuration file gets copied over to the installation
directory which is visible to users. So it is unnecessary to show a
FIXME message there. Move it to guc.c instead.

We will be tracking the actual fix of the FIXME separately.

(cherry picked from commit 32d33bf966bb3d55a98000c8a95b52a596ca6677)