Support DPE for duplicate sensitive random motions in Orca (#15629)
In certain cases, Orca generates duplicate sensitive motions, which in
the translator are interpreted as a filter on gp_segment_id. This
effectively only outputs the rows for a single segment. Inside the
optimizer, we treat this as a motion, as it causes the distribution to
change.

However, this isn't actually a physical motion, but rather a filter
doesn't create a separate slice. Therefore, we can pass partition
selectors through this, allowing Orca to generate more performant plans

For example, we can now generate plans such as
```
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..1293.01 rows=33 width=16)
   ->  Hash Join  (cost=0.00..1293.01 rows=11 width=16)
         Hash Cond: (part_mixed_dpe.b = non_part.b)
         ->  Append  (cost=0.00..862.00 rows=22 width=8)
               ->  Dynamic Seq Scan on part_mixed_dpe  (cost=0.00..431.00 rows=22 width=8)
                     Number of partitions to scan: 1 (out of 2)
               ->  Result  (cost=0.00..431.00 rows=22 width=8)
                     One-Time Filter: (gp_execution_segment() = 0)
                     ->  Dynamic Foreign Scan on part_mixed_dpe  (cost=0.00..431.00 rows=22 width=8)
                           Number of partitions to scan: 1 (out of 2)
         ->  Hash  (cost=431.00..431.00 rows=5 width=8)
               ->  Partition Selector (selector id: $0)  (cost=0.00..431.00 rows=5 width=8)
                     ->  Broadcast Motion 3:3  (slice2; segments: 3)  (cost=0.00..431.00 rows=5 width=8)
                           ->  Seq Scan on non_part  (cost=0.00..431.00 rows=2 width=8)
 Optimizer: Pivotal Optimizer (GPORCA)
(15 rows)
```

(cherry picked from commit 4484f5f1a43c0e3a780d6e5176edfa8e9ad96146)