ORCA: update relcache logic for setting hashable for RANGETYPES
Previously, ORCA would set if a datatype is hashable or not based on the
equality operator. As part of postgres 9.4 merge commit
(https://github.com/postgres/postgres/commit/36ea99c84d856177ec307307788a279cc600566e)
range's are treated as containers and whether a range type can support
hashing is decided based on hashing support of the range's subtype. This
check is done as part of fetching HASH_PROC in lookup_typ_cache(), which
is also later fetched when executing the query with a range type.
For a query where the range's subtype is not hashable, based on just the
operator's hashing support ORCA treats it as hashable and generates a
HASH plan which later fails during execution.  Below is an example:
```
create type cashrange as range (subtype = money);
select '(2,5)'::cashrange except select '(5,6)'::cashrange;
```

To fix this issue, ORCA adds a special case for rangetypes when doing a
lookup and sets hashable information for the type based on the HASH_PROC
info fetched instead of the operator.

Additionally, if any cast is added on top of a Scalar, ORCA would consider
the cast's Hash compatibility. This works fine for most of the cases,
but for Ranges we need to consider the underlying type's hashability.

(cherry picked from commit cbc3663dd675d198cb3ebf2a215a4abc7d695b87)