starting segments in execute mode post recovery (#15599)
After recovery, the segment should start in execute mode.
Currently, gprecoverseg starts the segment in utility mode after the recovery.
With these changes make sure the segment is started in execute mode.
Added tests to confirm segments are in execute mode after recovery.
Currently testing if the segment is in utility mode by connecting to
individual segments and checking the error thrown.