[ORCA] Support BitmapIndex plans for ArrayCmp on Hash indexes
Previously, when we added support for Hash Indexes we only considered it
for Equality comparator and not for ScalarArrayCmp. As part of commit
327592f, we added support to consider Bitmap scans for OR and IN
predicates on btree indexes. Need to extend it to consider Bitmap scan
alternatives for ArrayCmp predicates on hash indexes too.
This commit also updates the var names to reflect that we consider
Bitmap alternatives for other indexes(btree and hash) and not just for
btree indexes.

(cherry picked from commit fe394b08caaeb9db8a5307330aebad5ed7c12e8b)