Fix the locus of foreign upper path (#15821)
When `mpp_execute` is set to 'all segments' and `num_segments` is set to a different number of the default segment number, then we need to generate a path with the correct number of locus and a correct distribution policy for foreign table.
But the legacy code fails to consider the FDW option 'num_segments', so this PR is to fix this issue.

Things this PR will do:
- Add `num_segments` option for postgres_fdw.
- Fix locus of foreign_upper_path.
- Add testing cases.

(cherry picked from commit 40b1175ff2199bd8b1a673b676acf7507282ac53)