Allow ORCA plans with empty target list (#15939)
Currently, in constructing the target list for a DML statement, we require the
target list to be not NULL. If the target list is NULL, we fall back in the
debug build. This PR changes this behavior by removing the assertion that the
target list has to be non-NULL. One example of an empty target list is inserting
rows without columns into tables without columns.

Co-authored-by: Jingyu Wang <wjingyu@vmware.com>
(cherry picked from commit dbc291a979c19c817b17bffba450c60c0a8eb328)