minirepro: collect using sequence
When a table has an explicitly created sequence, the sequence will
be ignored by pg_dump if it's not owned by the table.

In pg_dump, implicitly created sequence like SERIAL and GENERATED
AS IDENTITY, the referenced sequence will be dumped if the owning
table is dumped. But for explicitly created sequence, we should
specify sequence's oid if we want to dump the sequence object when
the sequence is not owned by table.

Collect oids of sequence which be referenced by table columns in
gp_dump_query_oids.

(cherry picked from commit 59ed52677672aa0c9bb2880214f65a06053fb34d)