Add error/warning when exchanging/attach an external table
In 7X, external tables are essentially just one kind of foreign tables. Also,
because EXCHANGE PARTITION is now just DETACH+ATTACH PARTITION, exchanging
partition with an external table has the same behavior as attaching a foreign
table in upstream.

Because of that, there were certain restrictions that got lost for exchanging
external tables: (1) we cannot exchange a writable external table in; (2) we
cannot exchange a readable external table in w/o "WITHOUT VALIDATION" clause.

For (1): we should bring that restriction back, because otherwise there would
always be an error when scanning that partition (unless the query is written in
a way that avoids scanning that particular partition). To disallow having a
write-only external partition makes more sense. And, keeping existing behavior
is probably preferred for those who still use external tables than foreign tables.

For (2): in 7X, WITH/WITHOUT VALIDATION becomes a no-op[1]: we will always
validate the new partition *if possible*, even user specifies WITHOUT VALIDATION.
In case it is not possible such as for foreign tables, we will ignore validation
which is the same behavior as ATTACH PARTITION in upstream [2]. Therefore, it
doesn't make sense to continue asking users to add a no-op clause anymore. The
ask for WITHOUT VALIDATION is essentially to raise user awareness about this
issue. Therefore, we can add WARNING to server the same purpose.

Adjusted a few tests accordingly.

[1] https://docs.vmware.com/en/VMware-Greenplum/7/greenplum-database/ref_guide-sql_commands-ALTER_TABLE.html
[2] https://www.postgresql.org/docs/12/sql-createforeigntable.html

(cherry picked from commit e19c85c144b1244b34a4fa20406559934f4db37e)