Add logic for checking the process holding the lock file exists
Issue:
Some utilities (gprecoverseg, gpstop, etc...) use a lock directory to prevent customers from running
multiple instances of the program simultaneously. The lock directory is created inside the
COORDINATOR_DATA_DIRECTORY at the start of program execution and should be removed when the execution completes.
However, the lock directory is not always removed if the user abruptly kills the program
during execution. Consequently, the next run of the same utility encounters an error, stating
that the lock directory already exists and asks the user to manually delete it, even if
the process holding the lock is already terminated.

Solution:
To address this issue, this commit introduces a modification in mainUtils.py.
When attempting to acquire the lock, the code now checks if the process ID in the PID file,
present in the lock directory, still exists. If the process has been killed, the script
automatically removes the lock directory and proceeds to acquire a new lock for the current run.
This ensures that the utility can continue its execution without being hindered by
orphaned lock files from previous runs.