Fix ORCA invalid processing of nested SubLinks referenced in GROUP BY clause.
The ORCA optimizer could fallback to the legacy optimizer due to incorrect
processing of nested SubLinks (SubLink contains one more SubLink or rtable
subquery inside) during query normalization.

The erroneous behaviour took place when a query had a nested SubLink in its
targetList and this SubLink was in GROUP BY clause itself (see the test
section).

During query normalization for queries like in the tests the
RunGroupingColMutator function is called. This function processed nested
SubLinks incorrectly. It has Query code branch, which is supposed to be
executed for various subqueries of the original query. However, for the
nested SubLink case, when mutating the subquery of the SubLink (at second query
level), the m_current_query_level field was not increased respectively,
because MutateQueryOrExpressionTree, which was responsible for calling the
RunGroupingColMutator on the lower subquery, did not know anything about the
context. Therefore, for the query like in the tests, wrong value of the
m_current_query_level led to an unnecessary modification of the Var's
varlevelsup value. And when Var's varlevelsup value had been unnecessary
modified, it started referencing different query level. This could lead to the
fallback or invalid plan construction at further planning stages.

This patch fixes erroneous behaviour of RunGroupingColMutator by replacing the
call of MutateQueryOrExpressionTree with the call of RunGroupingColMutator in
order to properly modify m_current_query_level when mutating lower level RTEs.

(cherry picked from commit 31f6eea7af48d80ee934e7c4d9c2296ed3dbed90)