Change Query Parameter fallback message in Orca to notice type (#16197)
Previously a fallback to planner for query parameters would be logged as an error instead of
a notice, which is misleading as this is an intentional fallback. Now,
it will display as a notice.

`THD000,ERROR,""GPDB Expression type: Query Parameter not supported in DXL"",`

will change to:

`THD000,NOTICE,""GPDB Expression type: Query Parameter not supported in DXL"",`

(cherry picked from commit cfa4f8073a78438dc1c4261768cb544eff2f22a2)