Fix Orca crash due to improper colref mapping with CTEs (#16212)
When deriving stats in Orca for CTEs, we need to remap the statistics so that
the CTE consumer uses the statistics of the CTE producer. This involves
remapping the colrefs of associated statistics objects. While we did this for
histograms and widths, we didn't do this for attnos, which then caused a crash
in ApplyCorrelatedStatsToScaleFactorFilterCalculation() as it wasn't able to
find the correct colref in the map.

Note that in the test case the error is expected (and Orca actually falls
back), but it shouldn't cause a crash. The behavior should be the same both
with and without extended statistics.

Fixes #15462

(cherry picked from commit 67a1bee4dcf405c336f5a1cd05aa4702f73e4886)