Remove a FIXME in tupser.c
The benefit from this optimization idea is plausible (basically, avoiding one
memcpy of the memtuple), but the changes incurred by this are just too
backwards and invasive for the effort to be worthy:

1. There is no memtuple to directly serialize from. The `PRIVATE_tts_memtuple`
field in `TupleTableSlot` is gone after the FIXME was added (d20e071e879).
Basically we do not store the memtuple outside of the AM layer anymore (like in
the executor memory). I think bringing `PRIVATE_tts_memtuple` back to
`TupleTableSlot` and utilize it everywhere is a clear backward change.

2. Assuming we somehow pass the memtuple into `SerializeTuple` as a function
argument. But on the deserialization side, we would still need to deserialize
it back to minimal tuple because all the deserilization logic and else utilizes
minimal tuple now. The only way to avoid that is to carry memtuple around and
use it, which is basically same as bringing `PRIVATE_tts_memtuple` back to
`TupleTableSlot`. We also might have to bring back the concept of
`GenericTuple`, and revert a lot of the changes that we have made to bring GPDB
aligned with the AM framework which is a big milestone in 7X. Not very
worthwhile IMO.

(cherry picked from commit 33e877344a6cf7d36c6f6ff468fd0b1bd9651273)