Allow utility mode on coordinator to not upgrade lock for SELECT locking clause
In 6ebce7333177 we allow that when the right conditions meet, we do not upgrade
lock for a SELECT statement.

This commit fixes a corner case with utility mode: when we are connecting the
coordinator in utility mode, and when we do not enable global deadlock detector
(GDD) but all the other conditions meet, we should not need to upgrade lock as
well because we are not going to acquire any locks on the segments, so we will
not cause any global deadlock. In fact utility mode on any segment seems won't
cause global deadlock, but we are making a conservative change here to meet our
current need: we can now allow many isolation tests to be run w/o GDD. Note
that we still need to turn off optimizer for them to be run because ORCA does
not emit LockRows plannode (see 0e1184414193), so set it in the setup phase of
all sessions in the tests that have SELECT ... FOR ... statements, and then
enable the tests.

Fix lockmode