Fix gpstop pipeline flakiness after #15727
Following the merging of changes from pull request #15727, there was an intermittent test failure in
gpstop. This situation occurred when testing how the gpstart process behaves in the presence of a
lock directory created by a previous gpstop operation.

In this specific scenario, the expected behavior was that the lock directory from gpstop shouldn't
disrupt the gpstart process. To examine this, a test was set up where a gpstop process was started
and stopped before completion. This created a lock directory in COORDINATOR_DATA_DIRECTORY.
Subsequently, an attempt was made to clean up any remaining parts of the postgres process on the
coordinator using the 'pkill postgres' command. This was essential to prevent issues where
gpstart might fail with errors like ‘Greenplum instance already exists’. But sometimes, the remaining
parts of the postgres process didn't get cleaned up properly, causing the test to fail with the error
‘gpstart error: Coordinator instance process running’.

To fix this, added a step for cleaning all the postgres processes on the host by removing the lock file
‘/tmp/.s.PGSQL.*' along with ‘pkill -9 postgres’ to make sure everything is cleaned up.

(cherry picked from commit 04647e24486337d3f307b88388d3df0a65ef578e)