Fix ORCA invalid processing of nested SubLinks under aggregates.
The ORCA optimizer could produce an invalid query plan and fallback to the
legacy optimizer due to incorrect processing of nested SubLinks (SubLink
contains one more SubLink or rtable subquery inside) during query normalization.

The issue could arise when query had a AggRef inside a Sublink and that
AggRef contained one more SubLink inside (see the example of the
query in the tests section).

While mutating the SubLink inside RunExtractAggregatesMutator, the Aggref branch
is executed for SubLink's query targetList, which mutates args of the AggRef.
And if inside this AggRef arg some Var had varlevelsup value greater than the
m_agg_levels_up, i.e the Var referenced a relation that is higher than the
AggRef, the Var was not mutated at all. The varlevelsup field was not modified
and because of that, when AggRef was pulled up to the zero query level (by
appending it to the context.m_lower_table_tlist), the Var, whose varlevelsup
was higher than AggRef's, and which remained unchanged, started referencing
the wrong relation. This could lead to the fallback or invalid plan construction
when processing attributes at further planning stages.

This patch makes the RunExtractAggregatesMutator modify the varlevelsup of the
Var relatively the AggRef level in order to preserve proper query level.

(cherry picked from commit c10ea53b69913e39f5b976e7239febb6e25d0795)